+++ docs/api.md
@@ -706,10 +706,11 @@ https://api-accounts.dev.lcip.org/v1/session/status \
 
 ### Response
 
-Successful requests will produce a "200 OK" response with the account uid in the JSON body object:
+Successful requests will produce a "200 OK" response with the details about the token in the JSON body object:
 
 ```json
 {
+  "state": "verified",
   "uid": "80dc2f2e373b4b3bb992468e6d578cd2"
 }
 ```
