+++ lib/tokens/session_token.js
@@ -4,7 +4,6 @@
 
 'use strict'
 
-const inherits = require('util').inherits
 const userAgent = require('../userAgent')
 const ONE_HOUR = 60 * 60 * 1000
 
@@ -16,105 +15,114 @@ const TOKEN_FRESHNESS_THRESHOLD = 50 * 365 * 24 * ONE_HOUR // 50 years or post Y
 module.exports = (log, Token, config) => {
   const features = require('../features')(config)
 
-  function SessionToken(keys, details) {
-    Token.call(this, keys, details)
-    this.setUserAgentInfo(details)
-    this.setDeviceInfo(details)
-    this.email = details.email || null
-    this.emailCode = details.emailCode || null
-    this.emailVerified = !! details.emailVerified
-    this.verifierSetAt = details.verifierSetAt
-    this.locale = details.locale || null
-    this.mustVerify = !! details.mustVerify || false
+  class SessionToken extends Token {
 
-    // Tokens are considered verified if no tokenVerificationId exists
-    this.tokenVerificationId = details.tokenVerificationId || null
-    this.tokenVerified = this.tokenVerificationId ? false : true
-  }
-  inherits(SessionToken, Token)
+    constructor(keys, details) {
+      super(keys, details)
+      this.setUserAgentInfo(details)
+      this.setDeviceInfo(details)
+      this.email = details.email || null
+      this.emailCode = details.emailCode || null
+      this.emailVerified = !! details.emailVerified
+      this.verifierSetAt = details.verifierSetAt
+      this.locale = details.locale || null
+      this.mustVerify = !! details.mustVerify || false
 
-  SessionToken.tokenTypeID = 'sessionToken'
-
-  SessionToken.create = function (details) {
-    details = details || {}
-    log.trace({ op: 'SessionToken.create', uid: details.uid })
-    return Token.createNewToken(SessionToken, details)
-  }
-
-  SessionToken.fromHex = function (string, details) {
-    log.trace({ op: 'SessionToken.fromHex' })
-    return Token.createTokenFromHexData(SessionToken, string, details || {})
-  }
-
-  SessionToken.prototype.lastAuthAt = function () {
-    return Math.floor(this.createdAt / 1000)
-  }
-
-  // Parse the user agent string, then check the result to see whether
-  // the session token needs updating.
-  //
-  // If the session token has not changed, allowing up to an hour of
-  // leeway for lastAccessTime, return false.
-  //
-  // Otherwise, update properties on this then return true.
-  //
-  // It is the caller's responsibility to update the database.
-  SessionToken.prototype.update = function (userAgentString) {
-    log.trace({ op: 'SessionToken.update', uid: this.uid })
-
-    var freshData = userAgent.call({
-      lastAccessTime: Date.now()
-    }, userAgentString, log)
-
-    if (this.isFresh(freshData)) {
-      return false
+      // Tokens are considered verified if no tokenVerificationId exists
+      this.tokenVerificationId = details.tokenVerificationId || null
+      this.tokenVerified = this.tokenVerificationId ? false : true
     }
 
-    this.setUserAgentInfo(freshData)
+    static create(details) {
+      details = details || {}
+      log.trace({ op: 'SessionToken.create', uid: details.uid })
+      return Token.createNewToken(SessionToken, details)
+    }
 
-    return true
+    static fromHex(string, details) {
+      log.trace({ op: 'SessionToken.fromHex' })
+      return Token.createTokenFromHexData(SessionToken, string, details || {})
+    }
+
+    lastAuthAt() {
+      return Math.floor(this.createdAt / 1000)
+    }
+
+    get state() {
+      if (this.tokenVerified) {
+        return 'verified'
+      } else {
+        return 'unverified'
+      }
+    }
+
+    // Parse the user agent string, then check the result to see whether
+    // the session token needs updating.
+    //
+    // If the session token has not changed, allowing up to an hour of
+    // leeway for lastAccessTime, return false.
+    //
+    // Otherwise, update properties on this then return true.
+    //
+    // It is the caller's responsibility to update the database.
+    update(userAgentString) {
+      log.trace({ op: 'SessionToken.update', uid: this.uid })
+
+      var freshData = userAgent.call({
+        lastAccessTime: Date.now()
+      }, userAgentString, log)
+
+      if (this.isFresh(freshData)) {
+        return false
+      }
+
+      this.setUserAgentInfo(freshData)
+
+      return true
+    }
+
+    isFresh(freshData) {
+      var result = this.uaBrowser === freshData.uaBrowser &&
+        this.uaBrowserVersion === freshData.uaBrowserVersion &&
+        this.uaOS === freshData.uaOS &&
+        this.uaOSVersion === freshData.uaOSVersion &&
+        this.uaDeviceType === freshData.uaDeviceType &&
+        (
+          ! features.isLastAccessTimeEnabledForUser(this.uid, this.email) ||
+          this.lastAccessTime + TOKEN_FRESHNESS_THRESHOLD > freshData.lastAccessTime
+        )
+
+      log.info({
+        op: 'SessionToken.isFresh',
+        uid: this.uid,
+        tokenAge: freshData.lastAccessTime - this.lastAccessTime,
+        fresh: result
+      })
+
+      return result
+    }
+
+    setUserAgentInfo(data) {
+      this.uaBrowser = data.uaBrowser
+      this.uaBrowserVersion = data.uaBrowserVersion
+      this.uaOS = data.uaOS
+      this.uaOSVersion = data.uaOSVersion
+      this.uaDeviceType = data.uaDeviceType
+      this.lastAccessTime = data.lastAccessTime
+    }
+
+    setDeviceInfo(data) {
+      this.deviceId = data.deviceId
+      this.deviceName = data.deviceName
+      this.deviceType = data.deviceType
+      this.deviceCreatedAt = data.deviceCreatedAt
+      this.callbackURL = data.callbackURL
+      this.callbackPublicKey = data.callbackPublicKey
+      this.callbackAuthKey = data.callbackAuthKey
+    }
   }
 
-  SessionToken.prototype.isFresh = function (freshData) {
-    var result = this.uaBrowser === freshData.uaBrowser &&
-      this.uaBrowserVersion === freshData.uaBrowserVersion &&
-      this.uaOS === freshData.uaOS &&
-      this.uaOSVersion === freshData.uaOSVersion &&
-      this.uaDeviceType === freshData.uaDeviceType &&
-      (
-        ! features.isLastAccessTimeEnabledForUser(this.uid, this.email) ||
-        this.lastAccessTime + TOKEN_FRESHNESS_THRESHOLD > freshData.lastAccessTime
-      )
-
-    log.info({
-      op: 'SessionToken.isFresh',
-      uid: this.uid,
-      tokenAge: freshData.lastAccessTime - this.lastAccessTime,
-      fresh: result
-    })
-
-    return result
-  }
-
-  SessionToken.prototype.setUserAgentInfo = function (data) {
-    this.uaBrowser = data.uaBrowser
-    this.uaBrowserVersion = data.uaBrowserVersion
-    this.uaOS = data.uaOS
-    this.uaOSVersion = data.uaOSVersion
-    this.uaDeviceType = data.uaDeviceType
-    this.lastAccessTime = data.lastAccessTime
-  }
-
-  SessionToken.prototype.setDeviceInfo = function (data) {
-    this.deviceId = data.deviceId
-    this.deviceName = data.deviceName
-    this.deviceType = data.deviceType
-    this.deviceCreatedAt = data.deviceCreatedAt
-    this.callbackURL = data.callbackURL
-    this.callbackPublicKey = data.callbackPublicKey
-    this.callbackAuthKey = data.callbackAuthKey
-  }
-
+  SessionToken.tokenTypeID = 'sessionToken'
   return SessionToken
 }
